One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/pl-merch/pl-merch-row/pl-merch-row.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { PlMerchRowComponent } from './pl-merch-row.component';

describe('Component: PlMerchRow', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [PlMerchRowComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([PlMerchRowComponent],
      (component: PlMerchRowComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(PlMerchRowComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(PlMerchRowComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-pl-merch-row></app-pl-merch-row>
  `,
  directives: [PlMerchRowComponent]
})
class PlMerchRowComponentTestController {
}

